requirements.txt
numpy~=1.19.5
opencv-python~=4.2.0.34
tensorflow~=1.14.0
Pillow~=5.1.0
torch~=1.5.1
openpifpaf~=0.10.0
dlib~=19.21.1